# docs/swagger.yaml
basePath: /
definitions:
...
  basis_internal_dto.CreateTeamRequest:
    properties:
      name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - name
    type: object
  basis_internal_dto.InviteMemberRequest:
    properties:
      role:
        enum:
        - admin
        - member
        type: string
      user_id:
        type: integer
    required:
    - role
    - user_id
    type: object
...
  basis_internal_dto.TeamResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
...
  gin.H:
    additionalProperties: {}
    type: object
host: localhost:8080
info:
  contact:
    email: support@basis.local
    name: API Support
  description: REST API сервиса управления задачами и командной аналитики.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Task Management System API
  version: "1.0"
paths:
...
  /api/v1/teams:
    get:
      description: Возвращает список всех команд, в которых состоит аутентифицированный
        пользователь.
      produces:
      - application/json
      responses:
        "200":
          description: Список команд пользователя
          schema:
            items:
              $ref: '#/definitions/basis_internal_dto.TeamResponse'
            type: array
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Получение списка команд пользователя
      tags:
      - teams
    post:
      consumes:
      - application/json
      description: Создает новую команду и автоматически делает текущего пользователя
        ее владельцем (owner).
      parameters:
      - description: Данные для создания команды
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/basis_internal_dto.CreateTeamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Команда успешно создана
          schema:
            $ref: '#/definitions/basis_internal_dto.TeamResponse'
        "400":
          description: Невалидное тело запроса или ошибки валидации
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Создание новой команды
      tags:
      - teams
  /api/v1/teams/{id}/invite:
    post:
      consumes:
      - application/json
      description: Добавляет пользователя в команду. Выполнять операцию могут только
        участники с ролью owner или admin. Выдать роль owner через приглашение нельзя.
      parameters:
      - description: ID команды
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: Данные приглашения (user_id и роль)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/basis_internal_dto.InviteMemberRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Участник успешно добавлен в команду
        "400":
          description: Невалидный ID команды, попытка выдать роль owner или пользователь
            уже в команде
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав (пользователь не является owner или admin)
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Команда или приглашаемый пользователь не найдены
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Приглашение нового участника в команду
      tags:
      - teams
...
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer <JWT_TOKEN>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"